backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
alembic
redis
python-dotenv